working on trim, will be complete in tomorrow.

+++ app/src/main/java/com/nick/sampleffmpeg/ui/view/TitleTimeLayout.java
@@ -214,7 +214,7 @@ public class TitleTimeLayout extends RelativeLayout  implements View.OnTouchList
 
             //if touch start & end in short time, press event will happen.
             if (((SystemClock.elapsedRealtime() - dragStartTime) < SINGLE_TAP_MAX_TIME)) {
-                double seekVideoTime = ((double)me.getX() / (double)Constant.SP_PER_SECOND / parentActivity.getDisplayMetric().scaledDensity) * 1000;
+                float seekVideoTime = ((float)me.getX() / (float)Constant.SP_PER_SECOND / parentActivity.getDisplayMetric().scaledDensity) * 1000;
                 parentActivity.setCurrentSeekTime(seekVideoTime);
 
                 /**
